inspire/requirements.txt
requests
pygeometa
psycopg2-binary
python-dotenv
pycurl